backend/pyproject.toml
[tool.poetry]
name = "npc-chatter-backend"
version = "0.1.0"
description = "NPC Chatter backend"
authors = ["Dev <dev@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^2.3"
SQLAlchemy = "^2.0"
flask-migrate = "^4.0"
flask-jwt-extended = "^4.4"
Flask-Cors = "^4.0"
python-dotenv = "^1.0"
psycopg2-binary = "^2.9"
Werkzeug = "^2.3"

[tool.poetry.dev-dependencies]
pytest = "^7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
